+++ game/bishop.py
@@ -4,10 +4,31 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from game.piece import Piece
 
 class Bishop(Piece):
-        
-        def __init__(self, color):
-                self.color = color
 
-        # Implementar la lógica de movimiento
-        def move(self, start_pos, end_pos):
-                raise NotImplementedError("This method should be overridden by subclasses")
+    def __init__(self, color):
+        self.color = color
+
+    def __str__(self):
+        return 'B' if self.color == "WHITE" else 'b'
+
+    # Implementar la lógica de movimiento
+    def move(self, start_pos, end_pos, board):
+        start_row, start_col = start_pos
+        end_row, end_col = end_pos
+
+        # Movimiento diagonal
+        if abs(start_row - end_row) == abs(start_col - end_col):
+            row_step = 1 if end_row > start_row else -1
+            col_step = 1 if end_col > start_col else -1
+
+            current_row, current_col = start_row + row_step, start_col + col_step
+            while current_row != end_row and current_col != end_col:
+                if board[current_row][current_col] is not None:
+                    return False
+                current_row += row_step
+                current_col += col_step
+
+            # Verificar si la casilla de destino está vacía o contiene una pieza del oponente
+            if board[end_row][end_col] is None or board[end_row][end_col].color != self.color:
+                return True
+        return False