+++ game/queen.py
@@ -9,6 +9,9 @@ class Queen(Piece):
     def __init__(self, color):
         self.color = color
 
+    def __str__(self):
+        return 'Q' if self.color == "WHITE" else 'q'
+
     # Implementar la lógica de movimiento
     def move(self, start_pos, end_pos):
         raise NotImplementedError("This method should be overridden by subclasses")